+++ noodles-bed/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * bed/record: Implement `std::error::Error::source` for errors.
+
 ## 0.6.0 - 2022-11-18
 
 ### Added